move tests to be with their implementations
