+++ calm-hub-ui/src/visualizer/components/drawer/Drawer.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { Drawer } from '../visualizer/components/drawer/Drawer.js';
+import { Drawer } from './Drawer.js';
 
 describe('Drawer', () => {
     it('should render Drawer', () => {
